backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
opencv-python
transformers
torch
torchvision
